Update Create - remove Data file logic - add RestaurantRequestInterface in logic Update Data - clear file - remove createOrUpdateRestaurant method - remove buildRestaurant method Update di.xml - add new preference for RestaurantRequestInterface Add RestaurantRequest class - implement createOrUpdateRestaurant method - implement buildRestaurant method Add RestaurantRequestInterface.php Update Update.php - add RestaurantRequestInterface in logic
