+++ app/code/Stoyanov/Restaurants/Controller/Restaurant/Create.php
@@ -6,14 +6,14 @@ namespace Stoyanov\Restaurants\Controller\Restaurant;
 
 use Magento\Framework\App\Action\{Action, Context};
 use Magento\Framework\View\Result\PageFactory;
-use Stoyanov\Restaurants\Helper\Data;
+use \Stoyanov\Restaurants\Api\RestaurantRequestInterface;
 
 class Create extends Action
 {
     public function __construct(
-        Context                       $context,
-        private PageFactory           $pageFactory,
-        private Data                  $data,
+        Context $context,
+        private PageFactory $pageFactory,
+        private RestaurantRequestInterface $restaurantRequest
 
     ) {
         parent::__construct($context);
@@ -22,7 +22,7 @@ class Create extends Action
     public function execute()
     {
         if ($this->_request->isPost()) {
-            $response = $this->data->createOrUpdateRestaurant($this->_request->getParams());
+            $response = $this->restaurantRequest->createOrUpdateRestaurant($this->_request->getParams());
             if (!empty($response['entity_id'])) {
                 $this->messageManager->addSuccess(__('A new restaurant is created!'));
             }
